backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.23.5
pillow==10.0.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
ultralytics==8.0.145
google-generativeai==0.3.0